requirements.txt
flask
flask_request_params
autopep8
python-dotenv
mysql-connector-python